+++ docs/changes.rst
@@ -15,6 +15,7 @@ all releases are available on `Anaconda.org <https://anaconda.org/pytask/pytask>
 - :gh:`50` implements correct usage of singular and plural in collection logs.
 - :gh:`51` allows to invoke pytask through the Python interpreter with ``python -m
   pytask`` which will add the current path to ``sys.path``.
+- :gh:`53` changes the documentation to furo.
 
 
 0.0.10 - 2020-11-18